requirements.txt
streamlit>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
numpy>=1.21.0,<1.25.0
pillow>=9.0.0
opencv-python>=4.5.0
imgaug>=0.4.0
git+https://github.com/frgfm/torch-cam.git@main#egg=torchcam